annotation for max_delay fix

+++ python_async_function/2-measure_runtime.py
@@ -5,7 +5,7 @@ import time
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-async def measure_time(n: int, max_delay: float) -> float:
+async def measure_time(n: int, max_delay: int) -> float:
   """
   Measure the total execution time for the imported
   wait_n function and returns the total time
